refactor: extract Editor component
